backend/requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiolimiter==1.1.0
aiomysql==0.2.0
aioodbc==0.5.0
aiosignal==1.3.1
alabaster==0.7.16
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
APScheduler==3.11.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.1.0
Authlib==1.3.2
autocommand==2.2.2
babel==2.16.0
backports.tarfile==1.2.0
bcrypt==4.2.0
beanie==1.26.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
colour==0.1.5
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
coverage==7.6.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
deepl==1.20.0
Deprecated==1.2.15
distro==1.9.0
dnspython==2.6.1
docutils==0.20.1
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.105.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
gast==0.6.0
gevent==24.11.1
geventhttpclient==2.3.3
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imagesize==1.4.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
kiwisolver==1.4.5
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.96
lazy-model==0.2.0
libclang==18.1.1
limits==3.13.0
locust==2.32.5
macholib==1.16.3
mariadb==1.1.10
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.3.0
motor==3.5.1
msgpack==1.1.0
multidict==6.6.4
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.38.0
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
pip-check-reqs==2.5.3
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==5.29.1
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.8.0
pymongo==4.8.0
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.1.2
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.2.0
regex==2024.7.24
reportlab==4.4.3
requests==2.32.3
rich==13.7.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.2
tenacity==8.5.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typeguard==4.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.24.0.post1
validators==0.34.0
wasabi==1.1.2
wcwidth==0.2.13
weaviate-client==3.25.3
websockets==12.0
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==7.2
